Add files via upload
My issues is finding the reasons why a particular format of styling keeps re-occurring from one section to the others, including divs. This happens on the Font-weight, the positions which are ever centrally aligned, and the colour  which is ever picking this code here #e2c42c;. On color, it somewhat picks the header text box styles.